requirements.txt
pyautogui
pynput
pyyaml